Move extend primitives require to lib
Leasson learned: When setting the `path:` parameter to a gem
in the Gemfile, the `.gemspec` file is executed as part of loading
the gem. If no path is set, bundler does not execute the `.gemspec`
file, because it already pulled all the information it needed from
it at that point to create the Gemfile.lock and manage the
dependencies.

As I had the require for the `extend_primitives.rb` on the `.gemspec`
file, it only extended the primitives when using the `path:` parameter.

That lead me to double check everything around the code pushed to
RubyGems, my local code, the packaging of the gems, etc. An hour
of uncertainty seeing that the code was exactly the same locally
as the code coming from ruby gems.

It is fixed now.
Lesson Learned.