Make the gem opinionated on zeroes and double zeroes
This change make helps with solving problems of double zeroes.
It is opnionated because I decided that 1000 will be transformed
into `one thousand`, and not `10 hundred`, or `ten zero zero`, is
just what makes more sense to me. Others would probably have something
to say about the matter, but as of right now, I'm using the gem for
addresses so this works better for me.